test: separate stories for correction review/request
