+++ packages/client/src/v2-events/features/events/actions/correct/request/CorrectionRequest.stories.tsx
@@ -17,16 +17,11 @@ import { expect, waitFor, within, userEvent } from '@storybook/test'
 import { ActionType, TestUserRole } from '@opencrvs/commons/client'
 import { testDataGenerator } from '@client/tests/test-data-generators'
 import { useDrafts } from '@client/v2-events/features/drafts/useDrafts'
-import {
-  tennisClubMembershipEventDocument,
-  tennisClubMembershipEventWithCorrectionRequest
-} from '@client/v2-events/features/events/fixtures'
+import { tennisClubMembershipEventDocument } from '@client/v2-events/features/events/fixtures'
 import { ROUTES } from '@client/v2-events/routes'
 import { AppRouter } from '@client/v2-events/trpc'
 import { router } from './router'
 import * as Request from './index'
-
-const generator = testDataGenerator()
 
 const meta: Meta<typeof Request.Pages> = {
   title: 'CorrectionRequest',
@@ -111,7 +106,7 @@ export const ReviewWithParentFieldChanges: Story = {
         element: <Outlet />,
         children: [router]
       },
-      initialPath: ROUTES.V2.EVENTS.REVIEW_CORRECTION.REVIEW.buildPath({
+      initialPath: ROUTES.V2.EVENTS.REQUEST_CORRECTION.REVIEW.buildPath({
         eventId: tennisClubMembershipEventDocument.id
       })
     },
@@ -294,70 +289,5 @@ export const Summary: Story = {
       void expect(await canvas.findByText('Riku Rouvila')).toBeInTheDocument()
       void expect(await canvas.findByText('Max McLaren')).toBeInTheDocument()
     })
-  }
-}
-
-export const ReviewCorrection: Story = {
-  loaders: [
-    async () => {
-      window.localStorage.setItem(
-        'opencrvs',
-        generator.user.token.localRegistrar
-      )
-      // Intermittent failures starts to happen when global state gets out of whack.
-      // This is a workaround to ensure that the state is reset when similar tests are run in parallel.
-      await new Promise((resolve) => setTimeout(resolve, 50))
-    }
-  ],
-  parameters: {
-    reactRouter: {
-      router: {
-        path: '/',
-        element: <Outlet />,
-        children: [router]
-      },
-      initialPath: ROUTES.V2.EVENTS.REVIEW_CORRECTION.REVIEW.buildPath({
-        eventId: tennisClubMembershipEventWithCorrectionRequest.id
-      })
-    }
-  },
-
-  play: async ({ canvasElement }) => {
-    const canvas = within(canvasElement)
-
-    await waitFor(async () => {
-      await expect(
-        canvas.queryByRole('button', { name: 'Continue' })
-      ).toBeNull()
-    })
-
-    await waitFor(async () => {
-      await expect(
-        canvas.getByRole('button', { name: /approve/i })
-      ).toBeInTheDocument()
-    })
-
-    await waitFor(async () => {
-      await expect(
-        canvas.getByRole('button', { name: /reject/i })
-      ).toBeInTheDocument()
-    })
-
-    await expect(canvas.queryByText('Add attachment')).not.toBeInTheDocument()
-
-    await userEvent.click(canvas.getByRole('button', { name: /approve/i }))
-    await userEvent.click(canvas.getByTestId('close-dialog'))
-
-    await userEvent.click(canvas.getByRole('button', { name: /reject/i }))
-    const confirmButton = canvas.getByRole('button', { name: /confirm/i })
-    await expect(confirmButton).toBeDisabled()
-
-    const reasonInput = canvasElement.querySelector('#reject-correction-reason')
-    if (reasonInput) {
-      await userEvent.type(reasonInput, 'Missing legal documentation')
-      await expect(confirmButton).toBeEnabled()
-    }
-
-    await userEvent.click(canvas.getByTestId('close-dialog'))
   }
 }
